Fixed some security issues

+++ ContactProject/Controllers/ContactsController.cs
@@ -126,6 +126,10 @@ namespace ContactProject.Controllers
         public ActionResult DeleteConfirmed(int id)
         {
             Contact contact = db.Contacts.Find(id);
+            if(!IsUserContact(contact))
+            {
+                return HttpNotFound();
+            }
             db.Contacts.Remove(contact);
             db.SaveChanges();
             return RedirectToAction("Index");
